requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==2.1.1
sqlalchemy==2.0.34
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
yfinance==0.2.44
plotly==5.24.1
pydeck==0.9.1
statsmodels==0.14.3
scikit-learn==1.5.2
torch==2.4.1
shap==0.46.0
ruff==0.6.9
black==24.8.0
tenacity==9.0.0
# GCP Cloud SQL connector for Postgres (IAM auth, no JSON key required)
cloud-sql-python-connector[pg8000]==1.9.1
pg8000==1.31.2